+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 CHANGELOG
 =========
 
-2020-05-19 (0.0.0)
+2020-05-19 (0.0.2)
 ------------------
+- Add `get_creator_node_id` function and `EPRType` enum.
 
-Example changelog entry
-
+2020-05-19 (0.0.1)
+------------------
 - Created this package